Fix slow samtools for crams
I introduced a regression several months ago. The problem was that
we were using samtools_od which was a singularity container of
samtools wrapped with samtoolsHelper.sh, which essentially let you
pass in separate bam/bai files. There were specific problems with
systems like DNANexus. Recent version (1.10+) of samtools support
this functionality natively.

The problem was that samtoolsHelper.sh had the samtools REF_CACHE
hard-coded. The regression happened when I changed the way
samtools_od runs inside a tmp directory, which broke the REF_CACHE
and caused samtools to start downloading references from EBI for
crams. It was extra bad because the cache lived in the tmp
directory and so wasn't being shared.

This commit:

* Updates to samtools-1.11

* Sets REF_CACHE directly in the pipeline scripts, so it's obvious
  where it points.