Protect against undefined indexUrl
After removing samtools_od, which used to handle automatically
retrieving indexes, what we currently do is check in the shell
scripts if indexUrl is set. But JavaScript has been passing in
"undefined", which is still a valid string. So samtools tries to
fetch the index named "undefined". This commit sets undefined
indexUrl explicitly to the empty string so samtools assumes the
index isn't provided. Default samtools behavior in this case is to
download the index matching the bam/cram filename.