Decouple gru version from data directory version
Generally the data directory doesn't change very often, but we've
been keeping everything in sync for simplicity. In the interest
of making deploys easier/faster, this update allows using
a data directory from an older release. This required version is
hard coded and checked at runtime to verify the provided data
directory is considered compatible.