Fix tmp directories bug
In a recent release we changed all bash scripts to run with
-eo pipefail. This improved our error detection, but introduced
a bug. Many of our scripts create a temporary directory to write
files to while running, and delete that directory at the end of
the script. When scripts error early that code is never run. The
primary purpose of this commit is to move that responsibility from
the scripts to our main program. We also now run all scripts in a
tmp directory, so they can freely write to the current working
directory as necessary.

Also improved the order of the docker file somewhat, and in the
process switched to using non-versioned calls to tools like
samtools. This avoids the need for symlinks and lets us update
them all at once in the future.